Upstream PR: https://github.com/MinecraftForge/EventBus/pull/52